icmp_trojan_v1.0/requirements.txt
pyinstaller
pycryptodome
tinyaes